refactor: centralized authentication logic to ProtectedRoute
