+++ frontend/src/components/auth/ProtectedRoute.tsx
@@ -1,22 +1,40 @@
-import React, { useEffect } from 'react';
-import { Navigate } from 'react-router-dom';
+import React, { useState, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 interface ProtectedRouteProps {
   children: React.ReactNode;
 }
 
 const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
-  const isAuthenticated = localStorage.getItem('adminAuth') === 'true';
+  
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const navigate = useNavigate();
 
-  // Optional: additional authentication check or token validation logic
   useEffect(() => {
-    // This could validate a JWT token, check token expiration, etc.
-    // For demo purposes, we're just using a simple localStorage flag
-  }, []);
-
-  if (!isAuthenticated) {
-    // Redirect to login if not authenticated
-    return <Navigate to="/admin/login" replace />;
+    const token = localStorage.getItem('adminToken');
+    if (!token) {
+      navigate('/admin/login');
+      return;
+    }
+    fetch('/api/admin/verify', {
+      headers: { 'Authorization': `Bearer ${token}` }
+    })
+      .then(res => {
+        if (res.ok) {
+          setIsAuthenticated(true);
+        } else {
+          localStorage.removeItem('adminToken');
+          navigate('/admin/login');
+        }
+      })
+      .catch(() => {
+        localStorage.removeItem('adminToken');
+        navigate('/admin/login');
+      });
+  }, [navigate]);
+  
+  if(!isAuthenticated){
+    return null;
   }
 
   return <>{children}</>;
